refactor(console-ui): split user-console IIFE into per-feature modules (PR 2/N)
Completes the modularization started in PR 1. The 1559-line DOMContentLoaded
IIFE in user-console.ts is now a 49-line orchestrator that resolves DOM refs,
calls each feature's init(), and kicks off the initial data loads.

Module map (largest 307 lines, most ≤ 150):
- refs.ts          — singleton DOM refs populated once at startup
- user-types.ts    — shared interface definitions (kept separate from
                     types.ts which serves the operator console)
- state.ts         — extended with all previously closure-scoped mutable
                     state plus setActiveConversation()
- toast.ts         — showToast, copyMsg, copyBubble, code-block copy delegate
- citations.ts     — buildCitationsHtml, revealCitations, expand handlers
- contextWindow.ts — context indicator chip + compact button
- scrollFab.ts     — scroll-to-bottom FAB + scrollToBottom helper
- sidebar.ts       — nav, collapse, mobile open/close, touch swipe
- settings.ts      — theme, presets, persisted preferences
- thread.ts        — append user/system/error bubbles, pending assistant
- conversations.ts — list, history, context menu, rename modal, delete
- streaming.ts     — streamQuery + nonStreamQuery + sendQuery
- slash.ts         — autocomplete, command picker, submitSlashCommand
- attachments.ts   — file/web/KB chips + popovers
- input.ts         — textarea, send button, slash + picker keyboard nav

Behavior is preserved exactly. The few `window`-attached globals (copyMsg,
toggleCitation, etc.) that the bubble templates and inline onclick attributes
still reference are now installed by their owning module's init().
